Remove the tests.py and reorganise the directory structure
